Update install.sh
It appeared that stock version of youtube-dl installed by default on Raspbian, a python tool used by mpv player to decode videos, is really outdated. There is a bug with this tool while using it with nohup, bug which has been corrected from a very long time by the team in charge of youtube-dl. The solution is, according to the authors of the script, to remove the stock version and reinstall it from scratch. 
In order to make this jarvis plugin usable again, this update seems necessary. At the moment, the plugin is not able to read any music, precisely because of this outdated tool.